Frontend-Service/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==11.0.3

# NATS JetStream for OMS event subscription
nats-py==2.6.0

# Authentication and Security
PyJWT==2.8.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Type checking
mypy==1.7.1
types-PyJWT==1.7.1

# Code formatting
black==23.11.0
isort==5.12.0